Added win/lose screen, made decor items actually work

+++ Assets/Scripts/Game/Tiles/DecorTile.cs
@@ -7,7 +7,7 @@ public class DecorTile : TileObject
 {
 
     [Header("Decor Boosts")]
-    public float DecorBuff;
+    public int DecorBuff;
 
     public override void Initialize(Vector3Int position)
     {
